[FEAT] UserDefaultsWrapper, UserManager역할 분리
