+++ GraceLog/Source/Util/Manager/UserDefaults/UserDefaultKey.swift
@@ -1,0 +1,17 @@
+//
+//  UserDefaultKey.swift
+//  GraceLog
+//
+//  Created by 이건준 on 12/10/25.
+//
+
+import Foundation
+
+enum UserDefaultKey: String {
+    case id
+    case name
+    case nickname
+    case message
+    case email
+    case profileImageURL
+}
